add common config
